Feature/remove todo and cleanup

+++ Exercises/CheckIfItIsAStraightLine.cpp
@@ -10,7 +10,7 @@ public:
 		int y1 = coordinates[0][1];
 		int x2 = coordinates[1][0];
 		int y2 = coordinates[1][1];
-		for (int i = 2; i < coordinates.size(); i++)
+		for (size_t i = 2; i < coordinates.size(); i++)
 		{
 			int x = coordinates[i][0];
 			int y = coordinates[i][1];
